Fix bug introduced by d8a079741ba8c28075fb62c6cef8013534a41038
Link.name was changed to allow blank entries, but this this incompatible with
the requirement of that field being unique. Django will store empty-string the
the name field is left blank in a form, but two empty-string values would
collide as being unique conflicts.

The solution to this is to intercept empty-string values and convert them to
NULL before being saved in the DB.